+++ src/NetCoreForce.FunctionalTests/CRUDTests.cs
@@ -45,37 +45,71 @@ namespace NetCoreForce.FunctionalTests
             await client.UpdateRecord<SfAccount>(SfAccount.SObjectTypeName, account.Id, account);
 
             //get newly updated
-            SfAccount updatedAccount = await client.GetObjectById<SfAccount>(SfAccount.SObjectTypeName, newAccountId);
-            Assert.True(updatedAccount != null, "Failed to retrieve updated object");
-            Assert.Equal(description, updatedAccount.Description);
-
-            //create second new object for testing update multiple
-            SfAccount secondNewAccount = new SfAccount();
-            string secondAccountName = string.Format("Test Object {0}", Guid.NewGuid().ToString());
-            secondNewAccount.Name = secondAccountName;
-
-            CreateResponse secondCreateResp = await client.CreateRecord<SfAccount>(SfAccount.SObjectTypeName, secondNewAccount);
-
-            Assert.True(!string.IsNullOrEmpty(secondCreateResp.Id), "Failed to create second new object");
-
-            //test update multiple
-            string multipleDescription1 = string.Format("Test Description {0}", Guid.NewGuid().ToString());
-            string multipleDescription2 = string.Format("Test Description {0}", Guid.NewGuid().ToString());
-            newAccount.Description = multipleDescription1;
-            secondNewAccount.Description = multipleDescription2;
-            List<UpdateMultipleResponse> responses = await client.UpdateRecords(new List<object>() { newAccount, secondNewAccount }, true);
-            Assert.True(responses.All(r => r.Success), "Failed to update multiple objects");
-
-            //get newly updated objects
-            string secondNewAccountId = secondCreateResp.Id;
-            updatedAccount = await client.GetObjectById<SfAccount>(SfAccount.SObjectTypeName, newAccountId);
-            SfAccount secondUpdatedAccount = await client.GetObjectById<SfAccount>(SfAccount.SObjectTypeName, secondNewAccountId);
-            Assert.True(updatedAccount != null && secondUpdatedAccount != null, "Failed to retrieve multiple updated objects");
-            Assert.Equal(updatedAccount.Description, multipleDescription1);
-            Assert.Equal(secondUpdatedAccount.Description, multipleDescription2);
+            SfAccount udpatedAccount = await client.GetObjectById<SfAccount>(SfAccount.SObjectTypeName, newAccountId);
+            Assert.True(udpatedAccount != null, "Failed to retrieve udpated object");
+            Assert.Equal(description, udpatedAccount.Description);
 
             //delete
             await client.DeleteRecord(SfAccount.SObjectTypeName, newAccountId);
+
+            //use queryall to find deleted record
+        }
+
+        [Fact]
+        public async Task CreateAnUpdateMultiple()
+        {
+            ForceClient client = await forceClientFixture.GetForceClient();
+
+            //create new object
+            SfAccount firstAccount = new SfAccount() { };
+            string firstAccountName = string.Format("Test Object {0}", Guid.NewGuid().ToString());
+            firstAccount.Name = firstAccountName;
+
+            CreateResponse createResp = await client.CreateRecord<SfAccount>(SfAccount.SObjectTypeName, firstAccount);
+            string firstAccountId = createResp.Id;
+            Assert.True(!string.IsNullOrEmpty(createResp.Id), "Failed to create new object");
+
+            //get new object
+            firstAccount = await client.GetObjectById<SfAccount>(SfAccount.SObjectTypeName, firstAccountId);
+
+            //create second new object for testing update multiple
+            SfAccount secondAccount = new SfAccount();
+            string secondAccountName = string.Format("Test Object {0}", Guid.NewGuid().ToString());
+            secondAccount.Name = secondAccountName;
+
+            CreateResponse secondCreateResp = await client.CreateRecord<SfAccount>(SfAccount.SObjectTypeName, secondAccount);
+            string secondAccountId = secondCreateResp.Id;
+            Assert.True(!string.IsNullOrEmpty(secondCreateResp.Id), "Failed to create second new object");
+
+            //get new object
+            secondAccount = await client.GetObjectById<SfAccount>(SfAccount.SObjectTypeName, secondAccountId);
+
+            //test update multiple
+            string firstUpdatedDescription = string.Format("Test Description {0}", Guid.NewGuid().ToString());
+            string secondUpdatedDescription = string.Format("Test Description {0}", Guid.NewGuid().ToString());
+            firstAccount.Description = firstUpdatedDescription;
+            secondAccount.Description = secondUpdatedDescription;
+
+            try
+            {
+                List<UpdateMultipleResponse> responses = await client.UpdateRecords(new List<SObject>() { firstAccount, secondAccount }, true);
+                Assert.True(responses.All(r => r.Success), "Failed to update multiple objects");
+            }
+            catch (Exception ex)
+            {
+                throw ex;
+            }
+
+            //get newly updated objects
+            string secondNewAccountId = secondCreateResp.Id;
+            SfAccount firstUpdatedAccount = await client.GetObjectById<SfAccount>(SfAccount.SObjectTypeName, firstAccountId);
+            SfAccount secondUpdatedAccount = await client.GetObjectById<SfAccount>(SfAccount.SObjectTypeName, secondAccountId);
+            Assert.True(firstUpdatedAccount != null && secondUpdatedAccount != null, "Failed to retrieve multiple updated objects");
+            Assert.Equal(firstUpdatedDescription, firstUpdatedAccount.Description);
+            Assert.Equal(secondUpdatedDescription, secondUpdatedAccount.Description);
+
+            //delete
+            await client.DeleteRecord(SfAccount.SObjectTypeName, firstAccountId);
             await client.DeleteRecord(SfAccount.SObjectTypeName, secondNewAccountId);
 
             //use queryall to find deleted record