+++ CHANGELOG.md
@@ -2,9 +2,13 @@
 
 ### 2020-07-06 v2.7.0
 
-* Client: Added support for updating multiple objects + tests
+* Client: Added support for updating multiple objects
+* Client: Added backwards-compatible serialization contract resolver to allow inclusion of object ID on patch requests
 * Linq: Added support for filtering on booleans
 * Client/Linq: Added support for NetStandard2.1 (netcoreapp3.x)
+* Models: Update standard models to SF API v48
+* ModelGenerator: Fixed issue where API version in config file was being ignored, and default of v44 was used instead
+
 
 ### 2019-12-17 v2.6.0